+++ cfripper/rules/cloudformation_authentication.py
@@ -12,10 +12,12 @@ under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
-from ..model.rule import Rule
+__all__ = ["CloudFormationAuthenticationRule"]
+from cfripper.model.rule import Rule
 
 
 class CloudFormationAuthenticationRule(Rule):
+    """This rule checks for hardcoded credentials"""
 
     REASON = "Hardcoded credentials in {}"