+++ cfripper/rules/ebs_volume_has_sse.py
@@ -12,11 +12,15 @@ under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
-from ..model.enums import RuleMode
-from ..model.rule import Rule
+__all__ = ["EBSVolumeHasSSERule"]
+from cfripper.model.enums import RuleMode
+from cfripper.model.rule import Rule
 
 
 class EBSVolumeHasSSERule(Rule):
+    """
+    Check that server side encryption is enabled for all EBS volumes.
+    """
 
     REASON = "EBS volume {} should have server-side encryption enabled"
     RULE_MODE = RuleMode.MONITOR