+++ cfripper/rules/hardcoded_RDS_password.py
@@ -12,12 +12,16 @@ under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
+__all__ = ["HardcodedRDSPasswordRule"]
 from pycfmodel.model.parameter import Parameter
 
-from ..model.rule import Rule
+from cfripper.model.rule import Rule
 
 
 class HardcodedRDSPasswordRule(Rule):
+    """
+    This rule checks that RDS clusters and instances don't expose their passwords
+    """
 
     REASON_DEFAULT = "Default RDS {} password parameter (readable in plain-text) for {}."
     REASON_MISSING_NOECHO = "RDS {} password parameter missing NoEcho for {}."