+++ cfripper/rules/iam_managed_policy_wildcard_action.py
@@ -12,13 +12,17 @@ under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
+__all__ = ["IAMManagedPolicyWildcardActionRule"]
 from pycfmodel.model.resources.iam_managed_policy import IAMManagedPolicy
 
-from ..config.regex import REGEX_WILDCARD_POLICY_ACTION
-from ..model.rule import Rule
+from cfripper.config.regex import REGEX_WILDCARD_POLICY_ACTION
+from cfripper.model.rule import Rule
 
 
 class IAMManagedPolicyWildcardActionRule(Rule):
+    """
+    This rule checks for wildcards in IAM Managed policies
+    """
 
     REASON = "IAM managed policy {} should not allow * action"