+++ cfripper/rules/iam_roles.py
@@ -12,13 +12,22 @@ under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
+__all__ = [
+    "IAMRolesOverprivilegedRule",
+    "IAMRoleWildcardActionOnPermissionsPolicyRule",
+    "IAMRoleWildcardActionOnTrustPolicyRule",
+]
 from pycfmodel.model.resources.iam_role import IAMRole
 
-from ..config.regex import REGEX_IS_STAR, REGEX_WILDCARD_POLICY_ACTION
-from ..model.rule import Rule
+from cfripper.config.regex import REGEX_IS_STAR, REGEX_WILDCARD_POLICY_ACTION
+from cfripper.model.rule import Rule
 
 
 class IAMRolesOverprivilegedRule(Rule):
+    """
+    Rule that checks for wildcards in resources for a set of actions and restricts managed policies
+    """
+
     def invoke(self, cfmodel):
         for logical_id, resource in cfmodel.Resources.items():
             if isinstance(resource, IAMRole):
@@ -55,6 +64,9 @@ class IAMRolesOverprivilegedRule(Rule):
 
 
 class IAMRoleWildcardActionOnPermissionsPolicyRule(Rule):
+    """
+    Rule that checks for wildcards in actions in IAM role policies
+    """
 
     REASON = "IAM role {} should not allow * action on its permissions policy {}"
 
@@ -67,6 +79,9 @@ class IAMRoleWildcardActionOnPermissionsPolicyRule(Rule):
 
 
 class IAMRoleWildcardActionOnTrustPolicyRule(Rule):
+    """
+    Rule that checks for wildcards in actions in IAM role assume role policy documents
+    """
 
     REASON = "IAM role {} should not allow * action on its trust policy"