+++ cfripper/rules/privilege_escalation.py
@@ -12,12 +12,16 @@ under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
+__all__ = ["PrivilegeEscalationRule"]
 from pycfmodel.model.resources.iam_policy import IAMPolicy
 
-from ..model.rule import Rule
+from cfripper.model.rule import Rule
 
 
 class PrivilegeEscalationRule(Rule):
+    """
+    Rule that checks for actions that allow privilege escalation in IAM policies
+    """
 
     REASON = "{} has blacklisted IAM action {}"
     IAM_BLACKLIST = set(